[Fibonacci] Modify fibonacci calcute type to avoid overflow problem.

+++ src/main/java/cc/xpbootcamp/warmup/fibonacci/Fibonacci.java
@@ -1,12 +1,12 @@
 package cc.xpbootcamp.warmup.fibonacci;
 
 public class Fibonacci {
-    public int calculate(int position) {
+    public long calculate(int position) {
         if (position == 1 || position == 2)
             return position;
 
-        int first = 1, second = 1;
-        int result = 0;
+        long first = 1, second = 1;
+        long result = 0;
         for (int x = 3; x <= position; x++) {
             result = first + second;
             first = second;
